Remove tee
